+++ src/main/java/ti4/commands/explore/ExploreUse.java
@@ -1,0 +1,78 @@
+package ti4.commands.explore;
+
+import java.util.Map;
+
+import org.apache.commons.lang3.StringUtils;
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.commands.OptionMapping;
+import net.dv8tion.jda.api.interactions.commands.OptionType;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import ti4.generator.Mapper;
+import ti4.helpers.AliasHandler;
+import ti4.helpers.Constants;
+import ti4.helpers.Helper;
+import ti4.map.Game;
+import ti4.map.Planet;
+import ti4.map.Player;
+import ti4.map.Tile;
+import ti4.map.UnitHolder;
+import ti4.message.MessageHelper;
+
+public class ExploreUse extends ExploreSubcommandData {
+
+    public ExploreUse() {
+        super(Constants.USE, "Draw and activate an explore card from the deck or discard");
+        addOptions(idOption.setRequired(true).setAutoComplete(true));
+        addOptions(new OptionData(OptionType.STRING, Constants.PLANET, "Planet to explore").setRequired(false).setAutoComplete(true));
+        addOptions(new OptionData(OptionType.STRING, Constants.FACTION_COLOR, "Faction or Color").setAutoComplete(true));
+        addOptions(new OptionData(OptionType.BOOLEAN, Constants.FORCE, "True to force the draw, even if none are in the deck"));
+    }
+
+    @Override
+    public void execute(SlashCommandInteractionEvent event) {
+        Game game = getActiveGame();
+        String id = event.getOption(Constants.EXPLORE_CARD_ID, "", OptionMapping::getAsString);
+        boolean force = event.getOption(Constants.FORCE, false, OptionMapping::getAsBoolean);
+        id = StringUtils.substringBefore(id, " ");
+
+        if (!Mapper.isValidExplore(id)) {
+            MessageHelper.sendMessageToEventChannel(event, "Invalid ID specified: " + id);
+        }
+
+        if (force || game.pickExplore(id) != null) {
+            OptionMapping planetOption = event.getOption(Constants.PLANET);
+            String planetName = null;
+            if (planetOption != null) {
+                planetName = planetOption.getAsString();
+                planetName = AliasHandler.resolvePlanet(planetName);
+            }
+            Tile tile = null;
+            if (game.getPlanets().contains(planetName)) {
+                for (Tile tile_ : game.getTileMap().values()) {
+                    if (tile != null) {
+                        break;
+                    }
+                    for (Map.Entry<String, UnitHolder> unitHolderEntry : tile_.getUnitHolders().entrySet()) {
+                        if (unitHolderEntry.getValue() instanceof Planet && unitHolderEntry.getKey().equals(planetName)) {
+                            tile = tile_;
+                            break;
+                        }
+                    }
+                }
+                if (tile == null) {
+                    MessageHelper.sendMessageToEventChannel(event, "System not found that contains planet: " + planetName);
+                    return;
+                }
+            }
+            Player player = game.getPlayer(event.getUser().getId());
+            player = Helper.getGamePlayer(game, player, event, null);
+            String messageText = player.getRepresentation() + " used explore card: " + id;
+            if (force) messageText += "\nTHIS CARD WAS DRAWN FORCEFULLY (if the card wasn't in the deck, it was created from thin air)";
+            resolveExplore(event, id, tile, planetName, messageText, player, game);
+        } else {
+            MessageHelper.sendMessageToEventChannel(event, "Explore Card ID: `" + id + "` is not in the deck or discard pile.");
+        }
+    }
+
+}
